Update tests & minor fixes
